Working Commit
Added bot startup info message
Changed package information to match current project
Changed README project description
Added constants file to store constants
Loaded constants in ready.ts and added time counter for it
Shifted repeated types to an interface
Removed unused imports
Made sure that no one except bot author is able to reload or stop the bot
Shifted comments in reload file
Shortened code for server slash command
Finished code for user part of user slash command, pending GuildMember part
Added time to command logger
Shifted reloading command time counter which made it more accurate

Commit took 6.25 hrs

+++ src/commands/utils/stop.ts
@@ -6,8 +6,13 @@ module.exports = {
 		.setDescription('Stops the bot!')
 		.setDefaultMemberPermissions(PermissionFlagsBits.Administrator),
 	async execute(interaction: CommandInteraction): Promise<void> {
+		if ((await interaction.client.application.fetch()).owner != interaction.user) {
+			await interaction.reply(`Shut Down Failed\n\n\`You are not the owner of this bot.\``);
+			return;
+		}
 		await interaction.reply('Shutting Down!');
 		interaction.client.destroy();
 		console.info('Bot has shut down!');
+		process.exit(); // Bypasses command logger
 	},
 };
